# api/docs/swagger.yaml
basePath: /
definitions:
...
  api.ChallengeStatusResp:
    properties:
      name:
        example: Web Challenge
        type: string
      status:
        example: deployed
        type: string
      updated_at:
        example: "2018-12-31T22:20:08.948096189+05:30"
        type: string
    type: object
...
  api.HTTPPlainResp:
    properties:
      message:
        example: Messsage in response to your request
        type: string
    type: object
...
info:
  contact:
    email: contact@sdslabs.co.in
    name: SDSLabs
    url: https://chat.sdslabs.co
  description: Authenticated API for Beast CTF challenge deployment and competition
    services.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Beast API
  version: "0.2"
paths:
...
  /api/status/all:
    get:
      consumes:
      - application/json
      description: This returns the challenges in the status provided, along with
        their name and last updated time.
      parameters:
      - description: Status type to filter with, if none specified then all
        in: query
        name: filter
        required: true
        type: string
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.ChallengeStatusResp'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPPlainResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPPlainResp'
      summary: Returns challenge deployment status from the beast database for the
        challenges which matches the stauts according to filter.
      tags:
      - status
  /api/status/all/{filter}:
    get:
      consumes:
      - application/json
      description: This returns the challenges in the status provided, along with
        their name and last updated time.
      parameters:
      - description: Status type to filter with, if none specified then all
        in: query
        name: filter
        required: true
        type: string
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.ChallengeStatusResp'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPPlainResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPPlainResp'
      summary: Returns challenge deployment status from the beast database for the
        challenges which matches the stauts according to filter.
      tags:
      - status
  /api/status/challenge/{name}:
    get:
      consumes:
      - application/json
      description: Returns challenge deployment status from the beast database, for
        those challenges which are not present a status value NA is returned.
      parameters:
      - description: Name of the challenge
        in: query
        name: name
        required: true
        type: string
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ChallengeStatusResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPPlainResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPPlainResp'
      summary: Returns challenge deployment status from the beast database.
      tags:
      - status
  /api/submit/challenge:
    post:
      consumes:
      - application/json
      description: Returns success or error response based on the flag submitted.
        Also, the flag will not be submitted if it was previously submitted
      parameters:
      - description: Name of challenge
        in: formData
        name: chall_id
        required: true
        type: string
      - description: Flag for the challenge
        in: formData
        name: flag
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ChallengeStatusResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPPlainResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.HTTPPlainResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPPlainResp'
      summary: Verifies and creates an entry in the database for successful submission
        of flag for a challenge.
      tags:
      - Submit
...
schemes:
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"